Update mobility
The heading steer demo is functional. The robot will steer to the target
heading. However, tuning is required. If the robot starts at a heading that
is 180 degrees away from the target, it will enter a permanent turn. The robot
will turn to the target heading even if it is started at a heading that is
90 degrees away.